cryptohome: Get rid of MockVaultKeyset, clean up more VK
This completely devirtualizes VaultKeyset and gets rid of the mock
version. This is mostly straightforward, as most of the few tests that
actually use the mocks can be pretty easily replaced with mocks of the
underlying file accesses (e.g. mock out WriteFileAtomicDurable instead
of Save).

One exception is that there's a test that mocks out the decryption to
inject every kind of crypto error. This is very difficult to simulate,
however by just reducing the test to verifying one kind of crypto error
we can make much easier. I would argue that testing every kind of crypto
error isn't that much more useful, since that's mostly just testing the
crypto->mount error conversion functions, not VaultKeyset or
KeysetManagement.

I also included some cleanups of includes, and using declarations in the
files I was touching. As well as moving the VaultKeyset blob definition
out of the public header and into the implementation.

BUG=b:264313226
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I7d70c670319b5982bab43401d8c9193983821cff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5530124
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>