ippusb_bridge: Defer IPP-over-USB setup when screen is locked
ippusb_bridge is started by udev when a printer supporting IPP-over-USB
is connected to the USB port. Unfortunately, it fails when access to
USB is blocked by usbguard. It happens when the printer is connected
after user locked the screen. The goal of this CL is to detect this
kind of situation and defer the configuration of IPP-over-USB
connection till access to USB ports is restored.
Now, when access to USB is denied, ippusb_bridge waits for the signal
propagated on DBus informing that usbguard is deactivated. Then it
tries to repeat configuration of IPP-over-USB connection.

BUG=b:331433231
TEST=manually on octopus and a printer HP ENVY Inspire 7900

Change-Id: Id73d6433d694d579ed7f9a51bbf7c7ad19ef7f78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5528312
Tested-by: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Piotr Pawliczek <pawliczek@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>