cryptohome: add AuthFactorManager::RemoveMigratedFingerprintAuthFactors
A new api in AuthFactorManager that removes all fingerprint
auth factors that are marked as migrated. Backing files are deleted.
In-memory auth factor map is updated. The removal is done
by searching and deleting auth factors one by one.

The impl is not atomic all-or-nothing as any error in the removal
process aborts the remaining work. In practice it is not a problem since
we can always retry.

BUG=b:327048355
TEST=New unit tests.

Change-Id: I09258dfa6b6767db3b1d51cc7fbf1dfd11279695
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5530070
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Zi Lin <lziest@google.com>
Commit-Queue: Zi Lin <lziest@google.com>