cryptohome: Break a circular dependency
Came across this issue when trying to cut down on the size of the
VaultKeyset API.

This breaks a circular dependency between file system keyset and vault
keyset. The only reason for the dependency was so that both types could
support being constructed from one another, but we can do this in a
non-circular way by having one of the types support both directions.

In this case, I moved the support to vault keyset. Two reasons for this:
first, as VK is the more complex type with actual features and
functions, while FSK is just a simple data type, it seemed better to
have VK depend on FSK rather than the reverse. And second, because FSK
is still used in new-style USS based code that doesn't depend on the
older VKs at all, it seemed better to not require such code to also
implicitly depend on VK.

BUG=b:264313226
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I5ad2a5f7dc4f5fa16fc5a1d8f4cb3e2cac0202df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5530419
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>