installer: Fix shellcheck lints for external vars
In chromeos-install, fix all instances of "SC2154 (warning): <var> is
referenced but not assigned." Use the `:?` construct to make it an error
if the variable is actually not set. This should never be an error in
practice since the variables are imported from `write_gpt.sh`. (The
minios vars are an exception, as they were added in the v3 layout.)

BUG=b:340212443
TEST=cros lint installer/chromeos-install
TEST=copy script to reven DUT and run it

Change-Id: I32da18879fbe4bce6f51d7bb6dc49675894c27c2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5535153
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>