installer: Drop self-sudo from chromeos-install
Drop the code to re-launch chromeos-install with `sudo` if not running
as root, and just fail instead. Code that launches chromeos-install
should already be running it as root, and manual invocations can use
`sudo` directly if needed.

Dropping this means it doesn't need to be dealt with in the C++ification
of chromeos-install. Also fixed a shellcheck lint in the conditional:
"SC3028 (warning): In POSIX sh, UID is undefined".

BUG=b:319550148, b:340212443
TEST=cros lint installer/chromeos-install
TEST=run `chromeos-install` as root
TEST=run `sudo -u chronos chromeos-install`, fails as expected
TEST=run script via reven's install UI

Change-Id: Iaf9035dd6a9e9f8082ce3f027439e37d31b900fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5539008
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>