power: backlight_tool: Don't use current brightness to determine initial brightness
Currently backlight_tool uses the current brightness level as a
starting point to determine the initial brightness. As a result
the initial brightness returned by the tool may vary based on
the current brightness level, in particular on devices with ALS.
This is not desired, the initial brightness is supposed to
be a static value. Use 0 as 'current' brightness level when
calculating the initial brightness to obtain a deterministic
value that matches the one at boot time.

BUG=b:339517563
TEST=echo 1 > /var/lib/power_manager/has_ambient_light_sensor
  backlight_tool --set_brightness_percent=100
  backlight_tool --get_initial_brightness --force_battery --lux=150
    => ${value}

  backlight_tool --set_brightness_percent=60
  backlight_tool --get_initial_brightness --force_battery --lux=150
    => same as ${value} with --set_brightness_percent=100

Change-Id: Ia60131d20d26a105c29cfc3f8f072569001dd090
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5539068
Reviewed-by: Yanye Li <yanyeli@google.com>
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>