modemfwd: write unit tests against future API of ModemFlasher
When ModemFlasher is split, TryFlash will become the API of the new
task object, and the substeps will become public on ModemFlasher.
Write unit tests against the substeps to ensure coverage is not lost
while the API is split.

The FRIEND_TEST macros will be removed when the substeps are promoted
to public visibility.

BUG=b:315488486
TEST=unit tests

Change-Id: I7b727ed9219d945f595c642e85d70e5392f1de55
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5539890
Tested-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Daniel Winkler <danielwinkler@google.com>