shill: Reject LocalDevice creation for concurrency conflicts
Within WiFiProvider, reject interface creation requests whenever there
is a concurrency conflict. This shouldn't make any difference to
existing use cases, since we're only rejecting requests that would fail
in the kernel anyway. The main difference is that they're now being
rejected by Shill, so they should fail with ConcurrencyNotSupported,
instead of in the kernel.

NOTE: This only applies to LocalDevices, which are P2P and hotspot
devices. When creating a new LocalDevice, we'll check for conflicts
with any existing devices (including station WiFi devices). When
creating or enabling a station WiFi device, we still  don't do any
concurrency checks. This is because the handling of conflicts in the
WiFi device enablement flow is more complex and requires it's own
dedicated CL.

BUG=b:336535499
TEST=FEATURES=test emerge-brya shill
TEST=Requests are rejected with ConcurrencyNotSupported as expected:
gPaste: http://gpaste/5132516478681088

Change-Id: I949a69a97b236157e14217c20762a35e1f064306
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5560231
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>