installer: Disable shellcheck lint for lab_preserve_logs code
In chromeos-install, disable "SC2013 (info): To read lines rather than
words, pipe/redirect to a 'while read' loop" for the `lab_preserve_logs`
code. It's not easy to test changes to this code, so assume that it's
working as intended for now.

With this change, `cros lint chromeos-install` is clean.

BUG=b:340212443
TEST=cros lint installer/chromeos-install
TEST=copy script to reven DUT and run it

Change-Id: I42ceff11de976893d13d1608fe83009584e39d0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5538794
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>