init: startup: Simplify FS tuning
Since quota is always enabled, enabling quota on a FS where quota is
already applied is a no OP, do it unilaterally.

If the tune2fs call fails, it is silently ignored.
Future code is adding a call to fsck beforehand to improve the success
rate of the command.

BUG=b:172221451
TEST=Unit test.

Change-Id: Iea61c36a3980255d0f753ed1ab862461df7eca1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5588254
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>