installer: Drop self-sudo from chromeos-install-kernel
Drop the code to re-launch chromeos-install-kernel with `sudo` if not
running as root, and just fail instead.

Dropping this means it doesn't need to be dealt with in the C++ification
of chromeos-install-kernel. Also fixes a shellcheck lint in the
conditional: "SC3028 (warning): In POSIX sh, UID is undefined".

See also https://crrev.com/c/5539008 which made the same change to
chromeos-install.

BUG=b:340212443
TEST=cros lint chromeos-install-kernel

Change-Id: Ib3425d626c800bf9f299a5478f0f83ad0c409375
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5546017
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>