crash: Extract common parts of EfiCrash into PstoreCrash
Pstore supports multiple backends to store kernel crash records. EFI
vars are one such backend, but there's also a ram backend (historically
called ramoops) and block device backend. There may be more backends in
the future.

Regardless of the backend used, the pstore filesystem exposes records in
a common format where each file has a header indicating the type of
record (panic, oops, etc.) and the part of the record the file
corresponds to. For example, Panic#2 Part#4 indicates that the file is
for the fourth part of the second panic recorded in the backend. There
can only ever be one part with the ramoops pstore backend, but the EFI
pstore backend can have many parts split across different files because
each file has a 1:1 correspondence with an EFI var.

The EfiCrash class encapsulates a crash record (e.g. the first panic or
Panic#1) by figuring out how many files in the pstore filesystem
represent the part of a particular crash record and exposing functions
to load those parts into a buffer and remove the pstore files for a
crash record. Let's move that common functionality into a parent class
that EfiCrash inherits from so that we can introduce a ramoops class
that implements similar loading and removing behavior.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: If345034bfaaf8391de26620e3e1f42a7dd7eacbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567164
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>