secagentd: ignore network namespaced traffic
Ignore traffic going to or coming from a network interface that belongs
to a non-init network namespace. The init network namespace is defined
as having a net_cookie == 1.

Previously only the ifindex of the network interface was used to
determine whether traffic was external; this is a problem since the same
ifindex will be reused in different network namespaces.

BUG=b:332534411
TEST=tast tests pass

Change-Id: I004d5a094b0a027bb6d3fd41113a89de53416fcb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5556201
Reviewed-by: Ryan Borzello <rborzello@google.com>
Auto-Submit: Jason Ling <jasonling@chromium.org>
Commit-Queue: Jason Ling <jasonling@chromium.org>
Tested-by: Jason Ling <jasonling@chromium.org>