shill: Remove accessors of IPConfigs on Network
IPConfig is considered as an internal implementation of Network now.
Remove the getters and setters to avoid it being accessed outside the
Network class.

The only exception now is that Service can get the current IPConfig
object configured on Network. This will be removed eventually.

BUG=b:307854702
BUG=b:340974631
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: If0e5e7f8bcc0f423802afb7f91b54e3219cb75cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5547400
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>