shill: add the DHCPCDControllerInterface and the legacy implementation
We are planing to uprev dhcpcd from 7 to 10. Before uprev, we need a new
proxy interface for the dhcpcd process that handles the different
approaches to communicate old and new dhcpcd.

This CL adds a new interface DHCPCDControllerInterface for this purpose.
Also, the CL adds the implementation for the legacy dhcpcd 7, which
logic is the same as the current DHCPProvider, DHCPCDListener, and
DHCPCDProxy.

Note: the logic related to lease file is not implemented yet.

BUG=b:342068333
TEST=FEATURES=test emerge-test shill

Change-Id: Ib46d755175fd43ca761ad87d10b37aaa34159e40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5547183
Reviewed-by: Zikai Chen <chenzikai@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>