lorgnette: Update debug logging levels
PFUFS should be set to 40.  Also, turn on debugging for the SANE DLL
itself (in addition to all the other backends).

When I turned on the SANE DLL debug, it initially did not work.  The
SANE wrapper was getting created (in RegisterDBusObjectsAsync) before
the OnStartup function was being called.  Since the only thing OnStartup
did was set the logging level, I moved this to main and got rid of the
OnStartup stuff.

BUG=b:335543085
TEST=FEATURES=test emerge lorgnette

Change-Id: I7bac8622fe483fdd3495d485709e91a1a0adf7b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5547372
Commit-Queue: Nathan Muggli <nmuggli@google.com>
Tested-by: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Paul Moy <pmoy@chromium.org>