vm_tools: Fix maitred's invalid thread use.
D-Bus is initialized on the main thread, but actual D-Bus call
was made on the grpc thread, which is different from the main one.

Because, the shutdown is the only D-Bus call in the daemon,
this CL moves the initialization to inside the method.

BUG=b:341172347
TEST=Tryjob

Change-Id: Iba64e3c60e8b8deb5b4b1b3cc589c3df74b31fda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5550261
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Junichi Uekawa <uekawa@chromium.org>