crash: Generalize pstore crash collection
Pstore exports records in a similar way in the filesystem regardless of
the backend used to store those records. The first line of the file has
the type of record (oops, panic, etc.) and the part number (1, 2, 3,
etc.) and the rest of the file is the record data. With the efi-pstore
backend, the data is sometimes split across multiple parts because the
EFI backend isn't large enough to hold the entire panic or oops data for
a record.

With the ramoops backend, we don't have this complexity. Instead, the
ramoops backend stores the data in one part for one record. The common
part is the record header that indicates the type. Generalize that part
of the logic from CollectEfiCrashes() into the PstoreCrash class so that
it can be re-used by the ramoops-pstore backend collection code.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: Id5491c1ad292212b6c0432588e65e30a46c6b8bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567165
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>