crash: Move CrashSendingMode to its own file
Move the CrashSendingMode enum out of CrashCollector and into its own
file. Also make it clear that some of the structured metrics will be
using this enum.

This change will allow the metrics functions to be independent of
CrashCollector. (Needed since crash sender will be using them as well
and should not be linking in CrashCollector.)

BUG=b:177552411
TEST=CQ

Change-Id: I2b265832e39e77df3756e1f56747bb22afc0bddb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5551858
Reviewed-by: Hirthanan Subenderan <hirthanan@google.com>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Ian Barkley-Yeung <iby@chromium.org>