vm_tools: sommelier: Add quirk to fix viewport destination.
Fixes viewport destination to size of the screen for games using randr
emulation. For some games (like Barotrauma), xwayland sends the wrong
viewport destination, causing the game to render in a small box. This
adds a quirk to set the destination to the whole screen.

Not sure if the bug here is in xwayland or the game, but if it is
xwayland it should probably be fixed there at some point. This also
reproduces on other wayland compositors.

BUG=b:336844348
TEST=run barotrauma

Change-Id: Ie1e23c3b5c383a4c99dfbf10b2f4928b610a1253
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5555696
Commit-Queue: Arjun Srinivasan <asriniva@google.com>
Reviewed-by: Chloe Pelling <cpelling@google.com>
Tested-by: Arjun Srinivasan <asriniva@google.com>