crash: Remove pre 3.12 pstore record logic
A follow-up to CL:5247458.

We don't support kernels old enough to have this ramoops header.
Nowadays the header always indicates the type of record, e.g. Oops#1 is
the first oops or Panic#2 is the second panic.

Drop this record header as well. So far we've been including this header
in kcrash files when it isn't used. Stripping this header also makes the
logic match the efi-pstore code that already strips it.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I552809d34c1d9af0a8e1b15be5e87d99dfd5b748
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567158
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>