cryptohome: Fix a useless test on the status update timer.
The original test was testing that a AuthFactorStatusUpdate timer
correctly triggers. This test is currently not testing any thing
related to the update timer. At the time of its creation, it was
checking some AuthBlock function for intent lookup was invoked,
since that function should always run when the timer triggers.
Over time, the intent lookup gets refactored and the test stops
being useful. In fact it was just forwarding the time by 30 seconds
for nothing. Fix the test by using GetDelayInSeconds as the new
proxy check.

BUG=none
TEST=CQ

Change-Id: Ife49032c5fe03d18335dc2e22171cd3a9d01c897
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5555933
Tested-by: Zi Lin <lziest@google.com>
Commit-Queue: Zi Lin <lziest@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>