crash: Use RamoopsCrash class in CollectRamoopsCrashes()
Use the RamoopsCrash class in CollectRamoopsCrashes() now that all the
supporting pieces are in place. After this change we'll only collect
ramoops pstore records that are panics or corrupted. Otherwise they'll
be skipped. More importantly, the ramoops pstore records won't be
concatenated together, if for example there's an oops record and a panic
record for the oops we'll only collect the panic record which contains
the oops.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I591c0e84cfd9436dd243759a6e53ea756d79024f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567749
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>