fpstudy: Change pattern captures to PNM/simple
The ectool command and FPMCU firmware is designed to capture the pattern
samples in "simple" mode only. This "simple" mode means only returning
the image bytes offset inside of the fingerprint frame and automatically
converting to PNM. This assumption is baked into a few pieces of code
and documentation.

Although we should have support for returning the pattern as binary,
this would require a fair bit of refactoring over multiple changes.
The fact that we can request a "raw" binary version of pattern1 with
Elan's matcher is because their full frame size is equal to their
simple frame size (and FP_SENSOR_IMAGE_OFFSET = 0).

BUG=b:341944191
TEST=Capture on Gimble/Brya

Change-Id: I0ff6c83cd43ac61996553a3445d0f63fbc2ced93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5556423
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Craig Hesling <hesling@chromium.org>