crash: Return multiple statuses from CollectRamoops()
We're going to make CollectRamoops() return the status of multiple
crashes like CollectEfiCrashes() does. Make the return type into a
std::vector of CrashCollectionStatus enums and push through that change
to all the callers. Rename the function at the same time to reflect the
fact that multiple ramoops could be collected. We don't actually collect
more than one ramoops crash yet, but we will soon.

Add some more tests for WasKernelCrash() as well so that we know what
will happen if multiple ramoops crashes are collected (e.g. a corrupted
record and an oops).

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I4e07d2710651081dba62f7642bba9ebd6def5732
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567168
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Hirthanan Subenderan <hirthanan@google.com>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>