lorgnette: Update minijail to detect new devices
lorgnette is usually a short-running process, but there are certain
cases where it can run for an extended time.  In those cases, new USB
devices may get plugged in and would be expected to be discoverable by
the user.  Update the minijail config to allow these new devices to be
discovered.

BUG=b:335543085
TEST=run lorgnette, plug in device, and discover new devices

Change-Id: I7b7bc55fd58ca3a0c5d0cd13eb3b8af4a412167c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5564884
Tested-by: Nathan Muggli <nmuggli@google.com>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Nathan Muggli <nmuggli@google.com>