patchpanel: bug fix in lifelineFDService
When LifelineFDService automatically detects an invalid lifeline fd and
notifies the original caller that registered the fd, it is likely that
the caller will immediately trigger its unregistration callback, causing
a second call into DeleteLifelineFD onto the stack. This patch ensures
that this pattern is supported.

BUG=b:336883268
TEST=FEATURES=test emerge=$BOARD patchpanel

Change-Id: I6f0d29c5c0f0f127690fbf57243688cc6c97a3ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5568531
Tested-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Chih-Yu Huang <akahuang@chromium.org>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>