shill: Add queue for WiFi device requests
Add a struct type for tracking wifi device creation requests. These
structs contain the device type, the priority, and a callback to invoke
to actually create/enable the device.

Also adds a typedef for a set of these structs which are sorted by
priority.

BUG=b:328075299
TEST=FEATURES=test emerge-brya shill

Change-Id: Iaff0b3cc4a39f8f881ccf15796ed0ce931b69805
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5590132
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>