vm_tools: reduce balloon logs on high ram boards
This change increases the minimum size we need to change the balloon
before we will log a BalloonTrace message. The window size was
originally 100MB. This CL changes it to be 1/40 of RAM. This makes the
window size approximatley 100MB on 4GB machines, but proportionally
larger on higher memory machines.

BUG=b:342538043
TEST=tast run dut multivm.Lifecycle.crostini, verify balloon logs are
happening less.

Change-Id: I343a6ad5d7ae02dc4c0ad20ede81765b9692ae9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5573208
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Tested-by: Charles Dick <cwd@google.com>
Commit-Queue: Charles Dick <cwd@google.com>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>