patchpanel: Stop proxy NS and NA if ARC is non-interactive
This patch adds StartARCPacketFilter() and StopARCPacketFilter() to
GuestIPv6Service, which switches ndproxy for ARC devices between the
normal mode and a minimized mode. In the minimized mode, all NS and NA
packets are no longer forwarded to the downlink, and the host answers NS
on behalf of ARC for known global addresses to upstream neighbors
instead.

This is supposed to be used when ARC is in sleep mode, to kept if from
being waken by unrelated NS/NAs in the network, while keep it responsive
to existing connections.

BUG=b:320394039
TEST=manual on atlas: NotifyAndroidInteractiveState() and observe amount
     of icmp6 packets changes in ARC

Change-Id: Ic60bea043a27e91f368d9766b220d0d164372c2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5584440
Commit-Queue: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>