runtime_probe: change IPC init order
The destructor of |ScopedIPCSupport| will block the thread if there's
any pending mojo connectons haven't been disconnected. If the destructor
is called before any mojo remote in the same thread disconnect, there'll
be a deadlock that the mojo remote never disconnect and block forever.

This CL changes the init order by |ScopedIPCSupport|->|Context|, so
that the mojo remote in |Context| will always disconnect before
destructing |ScopedIPCSupport| when leaving the scope because the
destructors are called in the reversed order of the construction.

BUG=b:339912135
TEST=probe CPU 500 times and there's no deadlock

Change-Id: I66e72a882afb8f48772c8b1d92c3f669ce96a3c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5579317
Tested-by: Yu-An Wang <wyuang@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Yu-An Wang <wyuang@google.com>