shill: Add handling of 802.1x in UpdateKeyManagement
UpdateKeyManagement() was not meant initially to handle key management
for 802.1x.  Now this function is also called during SetWiFi() (because
key management might depend on the WPA3 support) so the case of 802.1x
is being signaled as unhandled.  Instead of silencing this log add
setting of the default key management for the 802.1x security class
- and make use of it where appropriate.

BUG=b:335021933
TEST=FEATURES=test emerge-$BOARD shill

Change-Id: Ibcdb227d00b24df2be74cb0568adb945ec34aef7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5580334
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Tested-by: Andrzej Ostruszka <andrzejo@google.com>