shill: Reland: Track WiFi device state changes in WiFiPhy
This change is a reland of the previously reverted
https://crrev.com/c/5560228. Testing confirms that ClaimDevice API
now works properly.

Currently, WiFiPhy only tracks the existence of WiFi device objects.
It should also track the UP/DOWN state of the objects, since the
UP/DOWN state is what matters for concurrency.

Accomplish this by having WiFi devices report state changed events to
WiFiProvider, which in turn reports the events to WiFiPhy. WiFiPhy is
responsible for probing the device state and determining whether the
device is UP accordingly.

BUG=b:314340771
TEST=FEATURES=test emerge-brya shill
TEST=dbus-send --system --fixed --print-reply --dest=org.chromium.flimflam / org.chromium.flimflam.Manager.ClaimDevice string:test string:wlan0

Change-Id: I24aa4fe4b50fb9beb1d1deff608d5c3f9aac2e52
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5623675
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>