crash: Return an enum from EfiCrash::GetType()
Clarify the record type strings we expect to find in the pstore
filesystem by returning an enum instead of a string from
EfiCrash::GetType(). This will make it easier to add a "corrupted" type
in the future because we won't need to make up a string to compare
against when the whole record is compressed and we don't know what type
of crash it is.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I515c2723eea2604ebd4bfc6222de9950a926d3b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5582073
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>