crash: Make EfiCrash copyable via assignment
The EfiCrash class isn't default copyable via assignment because it
stashes the KernelCollector that created it as a const reference. Make
that member into a pointer instead so this class can be copied via
assignment. Make the constructor and copy semantics explicit as well so
we know what the intended use of this class is.

BUG=b:330787908
TEST=cros_sdk cros_run_unit_tests --board=${BOARD} --packages crash-reporter

Change-Id: I2487463fef527c61c21a4a7c9f8c26f16b559ef4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5567163
Commit-Queue: Stephen Boyd <swboyd@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>