chromeos-config: Add _get_base_program function
Add a function to get the base program name. This will be used to
identify the base program mapped to a program and particularly used
when multiple programs are defined in program list and  use the build
packages as the base program. The function returns base program name if
it exist in program data otherwise returns the program name.

BUG=b:337986845
TEST=regen & run gen_config on fake program.

Cq-Depend: chromium:5575554
Change-Id: Ib8785c5f759eb03acf3bc4bcaaca56ea175a3e33
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5584043
Reviewed-by: Bob Moragues <moragues@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Aamir Bohra <aamirbohra@google.com>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Bhanu Prakash Maiya <bhanumaiya@google.com>
Commit-Queue: Aamir Bohra <aamirbohra@google.com>
Reviewed-by: YH Lin <yueherngl@chromium.org>