cros-disks: Mount /sys/devices/system/cpu/possible to drivefs minijail
A new change in drivefs has resulted in the system attempting to read
the /sys/devices/system/cpu/possible path to perform some CPU cache
operations. Given DriveFs is started in a minijailed environment it
doesn't have access. This results in it crashing on startup.

Once providing the required access, the next step was ensuring it didn't
crash due to the seccomp failure (required a new syscall for membarrier)
so permit access to that as well.

BUG=b:343807977
TEST=Deploy to betty and ensure ToT DriveFS starts

Change-Id: I1ed8bacccfa288543f4cc95feeb7ec03bb6a6b07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5581474
Reviewed-by: François Degros <fdegros@chromium.org>
Tested-by: Ben Reich <benreich@chromium.org>
Commit-Queue: Ben Reich <benreich@chromium.org>