fpstudy: Add groups discovery mechanism to analysis tool
This adds subcommand parsing to the analysis tool's run.py script.
The original functionality became the "report" subcommand
and the groups discovery mechanism became the "groups-discover"
subcommand.

The purpose of the groups-discover command is to scan the raw
fingerprint captures directory created by the study_serve
(fpstudy) tool to derive the user to group mapping. This
mapping should be saved in the User_groups.csv file, along
with the FAR/FRR matching decisions/outcomes. This allows
for running the analysis/report without the need for the
raw samples.

BUG=b:339298572
TEST=./run.py groups-discover ~/Documents/NoBackup/myfingers \
       ~/Documents/NoBackup/myfingers-outcomes/User_groups.csv
TEST=./run.py report ~/Documents/NoBackup/myfingers-outcomes analysis

Change-Id: I3ee6964cb47f61966a0f18dcbca861bd7cc8aa4f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5590290
Commit-Queue: Craig Hesling <hesling@chromium.org>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Tested-by: Craig Hesling <hesling@chromium.org>