shill: vpn: Parse remote traffic selectors for IKEv2
Parse the remote traffic selectors (remote_ts) from the output of
`swanctl --list-sas` and use it to generate the NetworkConfig for
routing setup. Note that if a packet has a dest ip which is not in the
remote_ts and this packet is sent to the xfrm interface, the packet will
just be dropped, so the behavior change of this change should always be
correct.

BUG=b:272229652
TEST=FEATURES="test" emerge-$BOARD shill
TEST=tast run $DUT network.VPNRouting.ikev2*

Change-Id: I3582f719e4322e1fcde874e3e984bf19476d766b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5587761
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>