Remove endlesspring@google.com from shadercached/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/endlesspring@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: I01ca200a6dda9e902625e1387803075a77b4c85b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5597619
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Commit-Queue: Renato Pereyra <renatopereyra@google.com>
Reviewed-by: Renato Pereyra <renatopereyra@google.com>
Tested-by: Renato Pereyra <renatopereyra@google.com>