Remove endlesspring@google.com from system_api/dbus/shadercached/OWNERS
This suggested change is automatically generated based on group
memberships and affiliations.
If this change is unnecessary or in error, vote CR -1 and the bot
will abandon it. Vote CR +1/2 to approve this change.

See the owner's recent activity for context:
https://chromium-review.googlesource.com/q/endlesspring@google.com

To report an issue, file a bug in the Infra>Codereview component.

Change-Id: Ia9503e7897ba05b9c70e6adb8a52a841dc3ad109
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5597559
Commit-Queue: Renato Pereyra <renatopereyra@google.com>
Reviewed-by: Renato Pereyra <renatopereyra@google.com>
Tested-by: Renato Pereyra <renatopereyra@google.com>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>