shill: Improve logic of Service::GetNetworkValidationMode()
Currently, when the CheckPortal property is set to automatic, the
ProxyConfig property has a higher priority than the technology-wide
control value (the CheckPortalList property on Manager) for deciding
the NetworkValidationMode. This patch flips the order of these two,
since the CheckPortalList property should be more reliable compared with
the inferred value from ProxyConfig.

The direct impact of this patch is that, with the current logic, a VPN
service with some ProxyConfig setting will have a NetworkValidationMode
value of HTTPonly, which will in turn put its service state to Connected
instead of Online, and cause some problem in routing. This patch fixes
this issue.

BUG=b:344798084
TEST=FEATURES="test" emerge-$BOARD shill
TEST=Set ProxyConfig on a VPN service, make sure that VPN routing is
  working well.

Change-Id: I7e5dc744bfb3d7645f5a205d0355b09190de990f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5599839
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>