flexor: Fix command execution
Currently, when executing a command through util::execute_command,
we call `Command::status`, which spawns a process and plugs Flexor's
stdout + stderr. In case we get a non-zero return code, we call
`Command::output`. This is wrong, as it doesn't return the output of
the previously executed command, but instead executes the command
again and collects stdout and stdin separately. This is wrong behaviour,
as we don't want to re-execute commands in error cases. Instead, we
can immediately call `Command::output` and print the stdout and stderr
in the error cases.

BUG=b:345125346
TEST=cros_run_unittests --board reven --package flexor

Change-Id: Iccd6cb729bfcb24bd199c424d8b6d39ea20f8a91
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5600618
Commit-Queue: Bastian Kersting <bkersting@google.com>
Reviewed-by: Thiago Bastos <tbastos@google.com>
Tested-by: Bastian Kersting <bkersting@google.com>